+++ subprojects/kotlinx-serialization/src/commonMain/kotlin/org/kotools/types/kotlinx/serialization/KotoolsTypesSerializers.kt
@@ -5,13 +5,11 @@ import kotlinx.serialization.modules.contextual
 import kotools.types.experimental.ExperimentalKotoolsTypesApi
 import kotools.types.internal.simpleNameOf
 import org.kotools.types.EmailAddress
-import org.kotools.types.Zero
 import org.kotools.types.internal.DeprecatedAsErrorSince
 import org.kotools.types.internal.ExperimentalSince
 import org.kotools.types.internal.KotoolsTypesVersion
 import org.kotools.types.internal.Warning
 import org.kotools.types.kotlinx.serialization.internal.EmailAddressAsStringSerializer
-import org.kotools.types.kotlinx.serialization.internal.ZeroAsByteSerializer
 
 /**
  * Contains modules for serializing the types provided by Kotools Types using
@@ -58,21 +56,6 @@ public object KotoolsTypesSerializers {
         get() = SerializersModule {
             contextual(EmailAddressAsStringSerializer)
         }
-
-    /** Returns the module for serializing the [Zero] type. */
-    @Deprecated(
-        "Use the 'ZeroAsByteSerializer' type instead.",
-        ReplaceWith(
-            "SerializersModule { contextual(ZeroAsByteSerializer()) }",
-            "kotlinx.serialization.modules.SerializersModule",
-            "kotlinx.serialization.modules.contextual",
-            "org.kotools.types.kotlinx.serialization.ZeroAsByteSerializer"
-        ),
-        DeprecationLevel.ERROR
-    )
-    @DeprecatedAsErrorSince(KotoolsTypesVersion.V4_5_3)
-    public val zero: SerializersModule
-        get() = SerializersModule { contextual(ZeroAsByteSerializer) }
 
     /**
      * Returns the string representation of this object.