+++ CHANGELOG.md
@@ -22,6 +22,11 @@ All notable changes to this project will be documented in this file.
   to 1.9, and kotlinx.serialization from 1.5.1 to 1.6.0 ([#407]).
 - Use default package names for Linux x64, macOS x64 and MinGW x64 ([#405]). 
 
+### 🔥 Removed
+
+- `KotoolsTypesSerializers.zero` **deprecated** property from the
+  **experimental** API ([#723]).
+
 ### 🐛 Fixed
 
 - Links to the API reference from the README documentation ([#742]).
@@ -34,6 +39,7 @@ Thanks to [@bodiam] and [@LVMVRQUXL] for contributing to this new release. 🙏
 [@LVMVRQUXL]: https://github.com/LVMVRQUXL
 [#405]: https://github.com/kotools/types/issues/405
 [#407]: https://github.com/kotools/types/issues/407
+[#723]: https://github.com/kotools/types/issues/723
 [#742]: https://github.com/kotools/types/pull/742
 
 ## 🔖 Releases